+++ Common/Extensions.cs
@@ -2,7 +2,7 @@
 using System.Drawing;
 using System.Windows.Forms;
 
-namespace Common
+namespace Alyn.Pointer.Common
 {
     public static class Extensions
     {
