+++ Player/Alyn.Pointer.App.csproj
@@ -9,8 +9,8 @@
     <ProjectGuid>{BF642637-49CE-469D-8B59-2EEED69234C2}</ProjectGuid>
     <OutputType>WinExe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Player</RootNamespace>
-    <AssemblyName>Player</AssemblyName>
+    <RootNamespace>Alyn.Pointer.App</RootNamespace>
+    <AssemblyName>Alyn.Pointer.App</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <FileUpgradeFlags>
@@ -44,8 +44,8 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
-    <Prefer32Bit>false</Prefer32Bit>
-    <PlatformTarget>x86</PlatformTarget>
+    <Prefer32Bit>true</Prefer32Bit>
+    <PlatformTarget>AnyCPU</PlatformTarget>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
@@ -212,25 +212,25 @@
     </BootstrapperPackage>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\YoloPipe.Harness\YoloPipe.Harness.csproj">
+    <ProjectReference Include="..\YoloPipe.Harness\Alyn.Pointer.YoloPipe.Harness.csproj">
       <Project>{94250f9a-51a0-4f51-b58e-35bde7dd1b43}</Project>
-      <Name>YoloPipe.Harness</Name>
+      <Name>Alyn.Pointer.YoloPipe.Harness</Name>
     </ProjectReference>
-    <ProjectReference Include="..\YoloPipe\YoloPipe.csproj">
+    <ProjectReference Include="..\YoloPipe\Alyn.Pointer.YoloPipe.csproj">
       <Project>{d2247033-9614-4e77-bb69-f4ea5aa5b884}</Project>
-      <Name>YoloPipe</Name>
+      <Name>Alyn.Pointer.YoloPipe</Name>
     </ProjectReference>
-    <ProjectReference Include="..\ObjectDetector\ObjectDetector.csproj">
+    <ProjectReference Include="..\ObjectDetector\Alyn.Pointer.ObjectDetector.csproj">
       <Project>{2522fe11-39c4-4981-a328-f5fe06fde235}</Project>
-      <Name>ObjectDetector</Name>
+      <Name>Alyn.Pointer.ObjectDetector</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Common\Common.csproj">
+    <ProjectReference Include="..\Common\Alyn.Pointer.Common.csproj">
       <Project>{462b50fc-50ba-4738-869e-dc6234fbd510}</Project>
-      <Name>Common</Name>
+      <Name>Alyn.Pointer.Common</Name>
     </ProjectReference>
-    <ProjectReference Include="..\TobiiAgent\TobiiAgent.csproj">
+    <ProjectReference Include="..\TobiiAgent\Alyn.Pointer.TobiiAgent.csproj">
       <Project>{9c8933a3-2caa-477b-a08b-41507683cfe7}</Project>
-      <Name>TobiiAgent</Name>
+      <Name>Alyn.Pointer.TobiiAgent</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
