+++ PointerAppAlyn.sln
@@ -3,17 +3,17 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29021.104
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TobiiAgent", "TobiiAgent\TobiiAgent.csproj", "{9C8933A3-2CAA-477B-A08B-41507683CFE7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Alyn.Pointer.TobiiAgent", "TobiiAgent\Alyn.Pointer.TobiiAgent.csproj", "{9C8933A3-2CAA-477B-A08B-41507683CFE7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YoloPipe", "YoloPipe\YoloPipe.csproj", "{D2247033-9614-4E77-BB69-F4EA5AA5B884}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Alyn.Pointer.YoloPipe", "YoloPipe\Alyn.Pointer.YoloPipe.csproj", "{D2247033-9614-4E77-BB69-F4EA5AA5B884}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YoloPipe.Harness", "YoloPipe.Harness\YoloPipe.Harness.csproj", "{94250F9A-51A0-4F51-B58E-35BDE7DD1B43}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Alyn.Pointer.YoloPipe.Harness", "YoloPipe.Harness\Alyn.Pointer.YoloPipe.Harness.csproj", "{94250F9A-51A0-4F51-B58E-35BDE7DD1B43}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UI", "Player\UI.csproj", "{BF642637-49CE-469D-8B59-2EEED69234C2}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Alyn.Pointer.App", "Player\Alyn.Pointer.App.csproj", "{BF642637-49CE-469D-8B59-2EEED69234C2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObjectDetector", "ObjectDetector\ObjectDetector.csproj", "{2522FE11-39C4-4981-A328-F5FE06FDE235}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Alyn.Pointer.ObjectDetector", "ObjectDetector\Alyn.Pointer.ObjectDetector.csproj", "{2522FE11-39C4-4981-A328-F5FE06FDE235}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "Common\Common.csproj", "{462B50FC-50BA-4738-869E-DC6234FBD510}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Alyn.Pointer.Common", "Common\Alyn.Pointer.Common.csproj", "{462B50FC-50BA-4738-869E-DC6234FBD510}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
