+++ YoloPipe/YoloPipe.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Threading.Tasks;
 
-namespace YoloPipe
+namespace Alyn.Pointer.YoloPipe
 {
     public class YoloPipe
     {
